Add files via upload
For the final capstone project of zero to hero I made a system that tracks inventory for a corner store. Products are stored as a class with attributes name, id, quantity (in stock) and price.
The system allows a user to make a sale, update inventory, add a new product to the inventory, display the total inventory, lookup a specific item in inventory, and close the store.
The items are stored on a object type 'store_inventory', but on a refactoring I would either give the object type more functionality or replace it entirely with the dictionary that makes the functions work in practice. Since all the functionality is available in the dictionary, I would chose that without a compelling reason to do otherwise